refactor: Q8 fixed-point exponential adaptation decay
Store adapt as i32 in Q8 fixed point so the geometric decay
adapt -= adapt >> adapt_decay stays exponential with time constant
~2^adapt_decay waves independent of magnitude. The integer right-shift
dead zone (adapt >> k == 0 while adapt < 2^k), which made small
adaptation ratchet up and permanently lock out weakly-driven neurons,
now sits below ~1/256 of a threshold unit, so adaptation always relaxes.
Effective threshold is threshold + (adapt >> ADAPT_SHIFT); bump saturates
at ADAPT_MAX. Potential leak is intentionally left unchanged: its low-end
dead zone enables sub-threshold integration of sparse input and cannot
lock out (firing resets it; negatives always relax).

+++ src/wave_net/network.rs
@@ -103,7 +103,8 @@ impl Network {
         self.layers[layer].lock().unwrap().potential[local]
     }
 
-    pub fn adaptation(&self, layer: usize, local: usize) -> i16 {
+    /// Raw Q8 fixed-point adaptation state (effective threshold contribution is `>> ADAPT_SHIFT`).
+    pub fn adaptation(&self, layer: usize, local: usize) -> i32 {
         self.layers[layer].lock().unwrap().adapt[local]
     }
 
@@ -233,7 +234,7 @@ mod tests {
             (0..net.layer_count())
                 .flat_map(|z| (0..(net.size() * net.size()) as usize).map(move |i| (z, i)))
                 .map(|(z, i)| net.adaptation(z, i))
-                .collect::<Vec<i16>>()
+                .collect::<Vec<i32>>()
         };
         assert_eq!(run(), run());
     }
